@@ -1,32 +1,15 @@
 import React from 'react';
-import { BrowserRouter as Router, useRoutes } from 'react-router-dom';
-import { Navbar } from './navbar';
-import { Footer } from './footer';
-import { Home } from './pages/home';
-import { People } from './pages/people';
-import { Planet } from './pages/planet';
-import { Starship } from './pages/starship';
-import { Film } from './pages/film';
-import { PEOPLE_URL, PLANET_URL, STARSHIP_URL } from './Constant';
+import { BrowserRouter as Router } from 'react-router-dom';
+import AppRoutes from './routes';
+import Navbar from './navbar/Navbar';
+import NavbarMobile from './navbar/NavbarMobile';
+import Footer from './footer/Footer';
 import './App.scss';
-import { Error } from './pages/error';
-
-const AppRoutes = () => {
-  const routes = useRoutes([
-    { path: '/', element: <Home /> },
-    { path: 'home', element: <Home /> },
-    { path: 'people', element: <People url={PEOPLE_URL} /> },
-    { path: 'planets', element: <Planet url={PLANET_URL} /> },
-    { path: 'starships', element: <Starship url={STARSHIP_URL} /> },
-    { path: 'films', element: <Film /> },
-    { path: '*', element: <Error /> },
-  ]);
-  return routes;
-};
 
 const App = () => (
   <Router>
     <Navbar />
+    <NavbarMobile />
     <div className="content">
       <AppRoutes />
     </div>
