@@ -1,6 +1,7 @@
 const { PUBLIC_URL } = process.env;
-export const PEOPLE_URL = 'https://swapi.dev/api/people';
-export const STARSHIP_URL = 'https://swapi.dev/api/starships';
-export const PLANET_URL = 'https://swapi.dev/api/planets';
+const API_BASE_URL = 'https://swapi.dev/api';
+export const PEOPLE_URL = `${API_BASE_URL}/people`;
+export const STARSHIP_URL = `${API_BASE_URL}/starships`;
+export const PLANET_URL = `${API_BASE_URL}/planets`;
 export const HOME_IMAGE_URL = `${PUBLIC_URL}/img/pexels-anna-tarazevich-5939140.jpg`;
 export const menus = ['Home', 'People', 'Planets', 'Starships', 'Films'];
